# adapters/target-ems-scada-http/openapi/v1.json
{
  "components": {
    "schemas": {
      "AcceptedCommand": {
        "properties": {
          "request_id": {
            "type": "string"
          },
          "result": {
            "$ref": "/bridge/v1/schemas/v1.11/command-result.schema.json"
          },
          "status_url": {
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status_url",
          "result"
        ],
        "title": "AcceptedCommand",
        "type": "object"
      },
...
      "CommandRequest": {
        "additionalProperties": false,
        "description": "Untrusted command fields that an adapter may deserialize from a request payload.\n\nAuthentication and target-instance context are deliberately absent. Unknown fields and\noperation variants fail deserialization instead of being ignored or treated as extensions.",
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "$ref": "/bridge/v1/schemas/v1.0/command.schema.json#/$defs/CorrelationId"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional caller-supplied correlation identity."
          },
          "expires_at": {
            "$ref": "/bridge/v1/schemas/v1.0/station-snapshot.schema.json#/$defs/UtcTimestamp",
            "description": "UTC instant after which admission or dispatch must fail."
          },
          "operation": {
            "$ref": "/bridge/v1/schemas/v1.0/command.schema.json#/$defs/CommandOperation",
            "description": "Explicit typed operation and parameters."
          },
          "request_id": {
            "$ref": "/bridge/v1/schemas/v1.0/command.schema.json#/$defs/RequestId",
            "description": "Caller-supplied idempotency identity."
          },
          "resource": {
            "$ref": "/bridge/v1/schemas/v1.0/station-snapshot.schema.json#/$defs/ResourceRef",
            "description": "Canonical charging resource, distinct from any selected bridge target."
          }
        },
        "required": [
          "request_id",
          "resource",
          "operation",
          "expires_at"
        ],
        "title": "CommandRequest",
        "type": "object"
      },
...
      "IntegrationError": {
        "properties": {
          "error": {
            "enum": [
              "ems_scada_http.request_conflict",
              "ems_scada_http.command_policy_rejected",
              "ems_scada_http.command_unsupported",
              "ems_scada_http.command_busy",
              "ems_scada_http.payload_too_large",
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential",
              "ems_scada_http.permission_denied",
              "ems_scada_http.unknown_resource",
              "ems_scada_http.unsupported_operation",
              "ems_scada_http.capacity_exhausted",
              "ems_scada_http.invalid_request",
              "ems_scada_http.bridge_required",
              "ems_scada_http.resource_not_found",
              "ems_scada_http.cursor_expired",
              "ems_scada_http.expired",
              "ems_scada_http.operation_not_supported",
              "ems_scada_http.source_unavailable",
              "ems_scada_http.deadline_exceeded"
            ],
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "integrationBearer": {
        "description": "Opaque integration token from the configured credential file, scoped by role and canonical resource. Never a management credential.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "Universal OCPP Bridge EMS/SCADA HTTP API",
    "version": "1.0.0"
  },
...
  "openapi": "3.1.1",
  "paths": {
...
    "/bridge/v1/commands": {
      "post": {
        "description": "Only control operations supported by the resource are admitted. The schema preserves canonical operation variants; privileged kind=ocpp requests return 403. request_id is at most 256 bytes and cannot be '.' or '..'. Unknown fields, origin injection and unknown operation variants are rejected. Body bytes and admission concurrency are bounded by capabilities.",
        "operationId": "commands",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "start": {
                  "value": {
                    "correlation_id": "workflow-42",
                    "expires_at": "2026-09-01T14:01:00Z",
                    "operation": {
                      "kind": "start",
                      "parameters": {
                        "authorization_reference": "local-auth-9"
                      }
                    },
                    "request_id": "cmd-target-start-7",
                    "resource": {
                      "bridge_id": "bridge-berlin-1",
                      "native_protocol_reference": {
                        "connector_id": 2,
                        "evse_id": 1,
                        "protocol": "ocpp201"
                      },
                      "resource": {
                        "connector_id": "connector-2",
                        "evse_id": "evse-1",
                        "kind": "evse"
                      },
                      "station_id": "station-7"
                    }
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedCommand"
                }
              }
            },
            "description": "Durably admit or replay an idempotent command. Admission is separate from protocol response and observed effect."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IntegrationError"
                    },
                    {
                      "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                    }
                  ]
                }
              }
            },
            "description": "ems_scada_http.command_policy_rejected, ems_scada_http.invalid_request, ems_scada_http.bridge_required, ems_scada_http.cursor_expired",
            "x-uob-error-codes": [
              "ems_scada_http.command_policy_rejected",
              "ems_scada_http.invalid_request",
              "ems_scada_http.bridge_required",
              "ems_scada_http.cursor_expired"
            ]
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unauthenticated, ems_scada_http.invalid_credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge",
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            },
            "x-uob-error-codes": [
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential"
            ]
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IntegrationError"
                    },
                    {
                      "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                    }
                  ]
                }
              }
            },
            "description": "ems_scada_http.permission_denied",
            "x-uob-error-codes": [
              "ems_scada_http.permission_denied"
            ]
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unknown_resource, ems_scada_http.resource_not_found",
            "x-uob-error-codes": [
              "ems_scada_http.unknown_resource",
              "ems_scada_http.resource_not_found"
            ]
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unsupported_operation",
            "x-uob-error-codes": [
              "ems_scada_http.unsupported_operation"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IntegrationError"
                    },
                    {
                      "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                    }
                  ]
                }
              }
            },
            "description": "ems_scada_http.request_conflict",
            "x-uob-error-codes": [
              "ems_scada_http.request_conflict"
            ]
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IntegrationError"
                    },
                    {
                      "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                    }
                  ]
                }
              }
            },
            "description": "ems_scada_http.expired",
            "x-uob-error-codes": [
              "ems_scada_http.expired"
            ]
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.payload_too_large",
            "x-uob-error-codes": [
              "ems_scada_http.payload_too_large"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IntegrationError"
                    },
                    {
                      "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                    }
                  ]
                }
              }
            },
            "description": "ems_scada_http.command_unsupported",
            "x-uob-error-codes": [
              "ems_scada_http.command_unsupported"
            ]
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.command_busy",
            "x-uob-error-codes": [
              "ems_scada_http.command_busy"
            ]
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.operation_not_supported",
            "x-uob-error-codes": [
              "ems_scada_http.operation_not_supported"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.capacity_exhausted, ems_scada_http.source_unavailable",
            "x-uob-error-codes": [
              "ems_scada_http.capacity_exhausted",
              "ems_scada_http.source_unavailable"
            ]
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.deadline_exceeded",
            "x-uob-error-codes": [
              "ems_scada_http.deadline_exceeded"
            ]
          }
        },
        "summary": "Durably admit or replay an idempotent command. Admission is separate from protocol response and observed effect."
      }
    },
    "/bridge/v1/commands/{request_id}": {
      "get": {
        "operationId": "command_status",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/bridge/v1/schemas/v1.12/command-result.schema.json"
                }
              }
            },
            "description": "Read a scoped command result belonging to this target instance."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.invalid_request, ems_scada_http.bridge_required, ems_scada_http.cursor_expired",
            "x-uob-error-codes": [
              "ems_scada_http.invalid_request",
              "ems_scada_http.bridge_required",
              "ems_scada_http.cursor_expired"
            ]
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unauthenticated, ems_scada_http.invalid_credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge",
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            },
            "x-uob-error-codes": [
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential"
            ]
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.permission_denied",
            "x-uob-error-codes": [
              "ems_scada_http.permission_denied"
            ]
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unknown_resource, ems_scada_http.resource_not_found",
            "x-uob-error-codes": [
              "ems_scada_http.unknown_resource",
              "ems_scada_http.resource_not_found"
            ]
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unsupported_operation",
            "x-uob-error-codes": [
              "ems_scada_http.unsupported_operation"
            ]
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.expired",
            "x-uob-error-codes": [
              "ems_scada_http.expired"
            ]
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.payload_too_large",
            "x-uob-error-codes": [
              "ems_scada_http.payload_too_large"
            ]
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.operation_not_supported",
            "x-uob-error-codes": [
              "ems_scada_http.operation_not_supported"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.capacity_exhausted, ems_scada_http.source_unavailable",
            "x-uob-error-codes": [
              "ems_scada_http.capacity_exhausted",
              "ems_scada_http.source_unavailable"
            ]
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.deadline_exceeded",
            "x-uob-error-codes": [
              "ems_scada_http.deadline_exceeded"
            ]
          }
        },
        "summary": "Read a scoped command result belonging to this target instance."
      },
      "head": {
        "operationId": "command_status_head",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Read a scoped command result belonging to this target instance."
          },
          "400": {
            "description": "ems_scada_http.invalid_request, ems_scada_http.bridge_required, ems_scada_http.cursor_expired",
            "x-uob-error-codes": [
              "ems_scada_http.invalid_request",
              "ems_scada_http.bridge_required",
              "ems_scada_http.cursor_expired"
            ]
          },
          "401": {
            "description": "ems_scada_http.unauthenticated, ems_scada_http.invalid_credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge",
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            },
            "x-uob-error-codes": [
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential"
            ]
          },
          "403": {
            "description": "ems_scada_http.permission_denied",
            "x-uob-error-codes": [
              "ems_scada_http.permission_denied"
            ]
          },
          "404": {
            "description": "ems_scada_http.unknown_resource, ems_scada_http.resource_not_found",
            "x-uob-error-codes": [
              "ems_scada_http.unknown_resource",
              "ems_scada_http.resource_not_found"
            ]
          },
          "405": {
            "description": "ems_scada_http.unsupported_operation",
            "x-uob-error-codes": [
              "ems_scada_http.unsupported_operation"
            ]
          },
          "410": {
            "description": "ems_scada_http.expired",
            "x-uob-error-codes": [
              "ems_scada_http.expired"
            ]
          },
          "413": {
            "description": "ems_scada_http.payload_too_large",
            "x-uob-error-codes": [
              "ems_scada_http.payload_too_large"
            ]
          },
          "501": {
            "description": "ems_scada_http.operation_not_supported",
            "x-uob-error-codes": [
              "ems_scada_http.operation_not_supported"
            ]
          },
          "503": {
            "description": "ems_scada_http.capacity_exhausted, ems_scada_http.source_unavailable",
            "x-uob-error-codes": [
              "ems_scada_http.capacity_exhausted",
              "ems_scada_http.source_unavailable"
            ]
          },
          "504": {
            "description": "ems_scada_http.deadline_exceeded",
            "x-uob-error-codes": [
              "ems_scada_http.deadline_exceeded"
            ]
          }
        },
        "summary": "Same handler as GET; response body omitted."
      }
    },
...
    "/bridge/v1/openapi.json": {
      "get": {
        "description": "Anonymous access only when no credential file is configured on loopback. Otherwise integrationBearer is required.",
        "operationId": "openapi",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Read the versioned OpenAPI document."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unauthenticated, ems_scada_http.invalid_credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge",
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            },
            "x-uob-error-codes": [
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential"
            ]
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.unsupported_operation",
            "x-uob-error-codes": [
              "ems_scada_http.unsupported_operation"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                }
              }
            },
            "description": "ems_scada_http.capacity_exhausted",
            "x-uob-error-codes": [
              "ems_scada_http.capacity_exhausted"
            ]
          }
        },
        "security": [
          {
            "integrationBearer": []
          },
          {}
        ],
        "summary": "Read the versioned OpenAPI document."
      },
      "head": {
        "description": "Anonymous access only when no credential file is configured on loopback. Otherwise integrationBearer is required.",
        "operationId": "openapi_head",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Read the versioned OpenAPI document."
          },
          "401": {
            "description": "ems_scada_http.unauthenticated, ems_scada_http.invalid_credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge",
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            },
            "x-uob-error-codes": [
              "ems_scada_http.unauthenticated",
              "ems_scada_http.invalid_credential"
            ]
          },
          "405": {
            "description": "ems_scada_http.unsupported_operation",
            "x-uob-error-codes": [
              "ems_scada_http.unsupported_operation"
            ]
          },
          "503": {
            "description": "ems_scada_http.capacity_exhausted",
            "x-uob-error-codes": [
              "ems_scada_http.capacity_exhausted"
            ]
          }
        },
        "security": [
          {
            "integrationBearer": []
          },
          {}
        ],
        "summary": "Same handler as GET; response body omitted."
      }
    },
...
  },
  "security": [
    {
      "integrationBearer": []
    }
  ],
  "servers": [
    {
      "description": "Selected EMS/SCADA integration listener; HTTPS outside isolated demos.",
      "url": "/"
    }
  ],
...
}